Bubble sort issue
Your release is sorting a already sorted array as it was unsorted, just putting a bool and changing the for to while we can get only one verification if the array is already sorted. In the cocktail there's not issue, but it verify 2x too, one more than necessary, not big deal edit.